refactor(core): move tests and benchmark outside of the src dir
This will allow #187 to be merged
