+++ packages/core/tests/diamond.test.ts
@@ -1,4 +1,4 @@
-import { declareAction, declareAtom, getState, map, combine } from '../index'
+import { declareAction, declareAtom, getState, map, combine } from '../src/index'
 
 describe('diamond problem (declareAtom)', () => {
   test('display name', () => {
