+++ packages/core/tests/log.json
